chore(match): rename /match -> /matches
